==> scripts/1_Analysis_multi_sites_GLS.R <==
# ---run climwin and GLS for all sites ---####

# clear environment ####
rm(list = ls())

# load libraries ####
library(climwin)
library(lme4)
library(mgcv)
library(splines)
library(gridExtra)
library(snow)
library(MuMIn)
library(tidyverse)
library(raster)

# prepare parameters ####
## paths to data ####

path_to_climate_data <- "https://raw.githubusercontent.com/forestgeo/Climate/master/Climate_Data/CRU/CRU_corrected/"

path_to_climate_data_NM <- "C:/Users/HerrmannV/Dropbox (Smithsonian)/GitHub/EcoClimLab/ForestGEO_dendro/data/climate/NM/CRU_climate/CRU_corrected/" # *TO BE EDITED* because the dendro repo is private... I ave to give absolute path (tokens are changing all the time in Github with private repos.... so it would be a pain to have to change them everytime...)

# path_to_BCI_pre <- "https://raw.githubusercontent.com/forestgeo/Climate/master/Climate_Data/Met_Stations/BCI/El_Claro_precip_starting_1929/pre_BCI.csv"
path_to_BCI_wet <- "https://raw.githubusercontent.com/forestgeo/Climate/master/Climate_Data/Met_Stations/BCI/El_Claro_precip_starting_1929/wet_BCI.csv"

path_to_CO2 <- "https://raw.githubusercontent.com/forestgeo/Climate/master/Other_environmental_data/CO2_data/CO2_MOANA_NOAA_combined.csv"

climate_variables <- c( "pre", "wet",
                        "tmp", "tmn", "tmx", "pet"#,
                        # "dtr", "cld") # "frs", "vap"
                              )


sites.sitenames <- c(BCI = "Barro_Colorado_Island", 
                     CedarBreaks = "Utah_Forest_Dynamics_Plot",
                     HarvardForest = "Harvard_Forest",
                     LillyDickey = "Lilly_Dickey_Woods",
                     SCBI = "Smithsonian_Conservation_Biology_Institute",
                     ScottyCreek = "Scotty_Creek",
                     Zofin = "Zofin",
                     HKK = "Huai_Kha_Khaeng",
                     NewMexico = "New_Mexico",
                     # Nebraska = "Niobara",
                     Niobara = "Niobara",
                     Hansley = "Niobara"
                     )

## analysis parameters ####
reference_dates <- list(BCI = c(30, 12),
                        CedarBreaks = c(30, 8),
                        HarvardForest = c(30, 8),
                        HKK = c(30, 12),
                        LillyDickey = c(30, 8),
                        NewMexico = c(30, 8),
                        SCBI = c(30, 8),
                        ScottyCreek = c(30, 8),
                        Zofin= c(30, 8),
                        # Nebraska = c(30, 8),
                        Niobara = c(30, 8),
                        Hansley = c(30, 8)) # refday in slidingwin
window_ranges <- list(BCI = c(15, 0),
                      CedarBreaks = c(15, 0),
                      HarvardForest = c(15, 0),
                      HKK = c(15, 0),
                      LillyDickey = c(15, 0),
                      NewMexico = c(15, 0),
                      SCBI = c(15, 0),
                      ScottyCreek = c(15, 0),
                      Zofin = c(15, 0),
                      # Nebraska = c(15, 0),
                      Niobara = c(15, 0),
                      Hansley = c(15, 0))# range in slidingwin
# reference_date <- c(30, 8) # refday in slidingwin
# window_range <- c(15, 0) #range in slidingwin

clim_var_group <- list(water = c("pre", "wet", "SF"),
                       temperature = c("tmp", "tmn", "tmx", "pet")#,
                       # c("dtr", "cld", "pet")
) # see issue 14, PET is in both the TMP and DTR groups. If it comes out as the best in both groups (should always be for the same time frame), then there are only 2 candidate variables for the GLS

clim_gap_threshold <- 5 # 5%

## variable units ####
variables_units <- c(pre = "(mm mo-1)",
                     wet = "(days mo-1)", 
                     tmp = "(C)", 
                     tmn = "(C)", 
                     tmx = "(C)", 
                     pet = "(mm day-1)", 
                     dtr = "(C)", 
                     cld = "%",
                     dbh = "(cm)",
                     CO2 = "(ppm)",
                     SF = "(m3 s-1)",
                     Year = "")
## sites abbrevation ####
# give site abbrevationtion in paper
sites_abb <- list(BCI  = "BCNM",
                  HKK = "HKK",
                  NewMexico = "LT",
                  CedarBreaks = "CB",
                  SCBI = "SCBI",
                  LillyDickey = "LDW",
                  HarvardForest = "HF",
                  # Nebraska = "NE",
                  Niobara = "NIO",
                  Hansley = "NE",
                  Zofin = "ZOF",
                  ScottyCreek = "SC")

# load data ####
## climate data ####

for(clim_v in climate_variables) {
  assign(clim_v, 
         rbind(
           read.csv(paste0(path_to_climate_data, clim_v,  "_CRU_corrected_conservative.csv")), #forestGEO sites #.1901.2019-ForestGEO_sites-6-03.csv")), 
           read.csv(paste0(path_to_climate_data_NM, clim_v, "_CRU_corrected_conservative.csv")) # NM site
           )
  )
  
}

# BCI_pre <- read.csv(path_to_BCI_pre, stringsAsFactors = F)
BCI_wet <- read.csv(path_to_BCI_wet, stringsAsFactors = F)

NE_SF <- read.csv("https://raw.githubusercontent.com/forestgeo/Climate/master/Climate_Data/Met_Stations/Niobrara/Niobrara_river_streamflow.csv")

clim_gaps <- read.csv("https://raw.githubusercontent.com/forestgeo/Climate/master/Climate_Data/CRU/CRU_corrected/CRU_gaps_analysisall_sites.reps.csv")
clim_gaps <- clim_gaps[clim_gaps$start_climvar.class %in% climate_variables, ]

## CO2 data ####
CO2 <- read.csv(path_to_CO2)

## core data ####
all_Biol <- read.csv("https://raw.githubusercontent.com/EcoClimLab/ForestGEO_dendro/master/data_processed/all_site_cores.csv?token=AEWDCIL3Z43KZUED5DXTUITBPMYIM")

all_Biol <- split(all_Biol, all_Biol$site)

sites <- names(all_Biol)
sites <- sites[!sites %in% "Hansley"] # remove Hansley

# prepare data ####
## species colors ####
all_species_codes <- lapply(all_Biol, function(x) levels(droplevels(factor(x$species_code))))
all_species_colors <- lapply(all_species_codes, function(x) {
  a <- scales::hue_pal()(length(x))
  names(a) <- x
  a
})
# sapply(all_species_colors, scales::show_col)
# names(species_colors) <- all_species_codes
# species_colors["CAOVL"] <- species_colors["CAOV"] # make CAOV and CAOVL the smae color


## climate data ####
for(clim_v in climate_variables) {
  print(clim_v)
  x <- get(clim_v)
  
  ### subset for the sites we care about
  x <- droplevels(x[x$sites_sitename %in% sites.sitenames, ])
  
  ### replace rows that are all NA by -999 for CedarCreaks by -999, just because that will cause issues later otherwise, but don't worry, those vairablewill be ignored later
  if(all(is.na(x[x$sites_sitename %in% sites.sitenames[c("CedarBreaks", "NewMexico")], -1]))) x[x$sites_sitename %in% sites.sitenames[c("CedarBreaks", "NewMexico")], -1] <- -999
  
  ### reshape to long format
  x_long <- reshape(x, 
                    times = names(x)[-1], timevar = "Date",
                    varying = list(names(x)[-1]), direction = "long", v.names = clim_v)
  
  ### format date
  x_long$Date <- gsub("X", "", x_long$Date)
  x_long$Date <- as.Date(x_long$Date , format = "%Y_%m_%d")
  
  
  ### combine all variables in one
  if(clim_v == climate_variables[1]) all_Clim <- x_long[, c(1:3)]
  else all_Clim <- merge(all_Clim, x_long[, c(1:3)], by = c("sites_sitename", "Date"), all = T)
  
}

### replace BCI wet by local data ####
idx_BCI <- all_Clim$sites_sitename %in% "Barro_Colorado_Island"


all_Clim$wet[idx_BCI] <- BCI_wet$climvar.val[match(format(all_Clim$Date[idx_BCI], "%Y-%m"), format(as.Date(BCI_wet$Date), "%Y-%m"))]
# all_Clim$pre[idx_BCI] <- BCI_pre$climvar.val[match(format(all_Clim$Date[idx_BCI], "%Y-%m"), format(as.Date(BCI_pre$Date), "%Y-%m"))]


### remove BCI wet from clim gap as it is not relevant anymore
clim_gaps <- clim_gaps[!(clim_gaps$start_sites_sitename %in% "Barro_Colorado_Island" & clim_gaps$start_climvar.class %in% c("wet")), ]

### add stream flow at NE ####
unique(all_Clim$site)

all_Clim$SF <- -999

NE_SF$Date <- as.Date(paste(NE_SF$year, NE_SF$monthno, "15", sep = "-"))

idx_NE <- all_Clim$sites_sitename %in% "Niobara"

all_Clim[idx_NE, ]$SF <- NE_SF$meandis_m3s[match(substr(all_Clim$Date[idx_NE], 1, 7), substr(NE_SF$Date, 1, 7))]


# keep only complete rows (this will remove BCI data for years where we don't have pre data, same for NE SF)

all_Clim <- all_Clim[complete.cases(all_Clim), ]

### add columns for detrend climate variables ####
for(clim_v in climate_variables) {

  all_Clim$y <- all_Clim[, clim_v]
  all_Clim[paste(clim_v, "detrended", sep = "_")] <- gam(y ~ s(as.numeric(Date), by = factor(sites_sitename)), data = all_Clim)$residuals
  all_Clim$y <- NULL
  
}

### format date to dd/mm/yyyy ####
all_Clim$Date <- format(all_Clim$Date, "%d/%m/%Y") 

### add site column
all_Clim$site <- names(sites.sitenames)[match(all_Clim$sites_sitename, sites.sitenames) ]


### save prepared clim data 
write.csv(all_Clim, "processed_data/Climate_data_all_sites.csv", row.names = F)


## cores ####


for (site in sites) {
  Biol <- all_Biol[[site]]
  Clim <- droplevels(all_Clim[all_Clim$sites_sitename %in% sites.sitenames[site], ])
  
  ### format date to dd/mm/yyyy ####
  Biol$Date <- paste0("15/06/", Biol$Year) # dd/mm/yyyy setting up as june 15, ARBITRATY
  
  ## remove years that are before climate record (+ first few first months to be able to look at window before measurement) ####
  Biol <- Biol[Biol$Year >= min(as.numeric(substr(Clim$Date, 7, 10)))+  window_ranges[[site]][1]/12, ]
  
  ## remove years that are after climate record ####
  Biol <- Biol[Biol$Year <= max(as.numeric(substr(Clim$Date, 7, 10))), ]
  
  
  
  ## remove any cores that have less than 30 years ####
  Biol <- Biol[Biol$coreID %in% names(which(table(Biol$coreID)>= 30)), ]
  
  
  
  ## consider coreID as factor ####
  Biol$coreID <- factor(as.character(Biol$coreID))
  ## save back into Biol ####
  
  all_Biol[[site]] <- Biol
  
  
}

# summarise variable-month combinations gaps ####
adjusted_clim_gaps <- list()
variables_to_drop <- list() # these will automatically be dropped because no matter what time window is best, all momths have > 5% records missing
for(site in sites) {
  
  Biol <- all_Biol[[site]]
  Clim <- droplevels(all_Clim[all_Clim$sites_sitename %in% sites.sitenames[site], ])
  clim_gap <- clim_gaps[clim_gaps$start_sites_sitename %in% sites.sitenames[site], ]
  
  if(nrow(clim_gap) > 0) {
    Clim$Year <- as.numeric(sapply(strsplit(Clim$Date, "/"), "[[", 3))
  # if(min(Biol$Year) != 1903) stop("I did not code for this eventuallity")
  start_year <- min(min(Biol$Year), min(Clim$Year)) # taking the min because technically Clim year starts earlier to be able to look before first measurement
  end_year <- max(Biol$Year) # taking Biol because already trimed to max date.
  n_years <- end_year - start_year
  max_number_months_missing <- floor(n_years * clim_gap_threshold/100)

  # adjust the number of years missing when start and end Dates are before or after (resp) the start and end year of the analysis
  adjusted_clim_gap <- clim_gap
  idx_starts_before <- clim_gap$start_Date < start_year
  idx_ends_after <- clim_gap$end_Date > end_year
  
  adjusted_clim_gap[idx_starts_before, ]$nyears <- adjusted_clim_gap[idx_starts_before,  ]$nyears - (start_year - clim_gap$start_Date[idx_starts_before])
  
  adjusted_clim_gap[idx_ends_after, ]$nyears <- adjusted_clim_gap[idx_ends_after,  ]$nyears -(clim_gap$end_Date[idx_ends_after] - end_year) 
  
  # replace negative values by 0
  adjusted_clim_gap$nyears[adjusted_clim_gap$nyears <0] <- 0 #adjusted_clim_gap <- adjusted_clim_gap[adjusted_clim_gap$nyears > 0, ]
  
  # make month and variable factors so that they are filled with 0 missing years
  adjusted_clim_gap$month <- factor(adjusted_clim_gap$month, levels = 1:12)
  adjusted_clim_gap$start_climvar.class <- factor(adjusted_clim_gap$start_climvar.class)
  
  # sum number of nyears per variable-month combo
  adjusted_clim_gap <- aggregate(formula = nyears ~ start_climvar.class + month,
                                 data = adjusted_clim_gap, 
                                 FUN = sum,
                                 drop = F)
  
  adjusted_clim_gap$nyears[is.na(adjusted_clim_gap$nyears)] <- 0
  
  # get to % missing data
  adjusted_clim_gap$percent_misssing <- adjusted_clim_gap$nyears *100/ n_years
  
  # get if month passes threshold
  adjusted_clim_gap$remove <- adjusted_clim_gap$nyears > max_number_months_missing
  
  # change factors back to characters
  adjusted_clim_gap$month <- as.character(adjusted_clim_gap$month)
  adjusted_clim_gap$start_climvar.class <- as.character(adjusted_clim_gap$start_climvar.class)
  
  # save what variables shold be removed and new adjusted gaps data
  variables_to_drop[[site]] <- names(which(tapply(adjusted_clim_gap$remove, adjusted_clim_gap$start_climvar.class, sum)==12))
  
  adjusted_clim_gaps[[site]] <- adjusted_clim_gap
  
  } 
}


# add CO2 data to all_Biol ####
all_Biol <- lapply(all_Biol, function(Biol) {
  Biol$CO2 <- CO2$CO2_ppm[match(Biol$Year, CO2$year)] 
  return(Biol)
  })

## Run the Analysis ####
for(solution_to_global_trend in c("none", "detrend_climate", "old_records_only", "young_records_only")[]) {
  
  last_year_older_records = 1970
  
  if(solution_to_global_trend %in% "none") {
    detrend_climate = FALSE
    old_records_only = FALSE
    young_records_only = FALSE
  }
  if(solution_to_global_trend %in% "detrend_climate") {
    detrend_climate = TRUE
    old_records_only = FALSE
    young_records_only = FALSE
  }
  if(solution_to_global_trend %in% "old_records_only") {
    detrend_climate = FALSE
    old_records_only = TRUE
    young_records_only = FALSE
  }
  if(solution_to_global_trend %in% "young_records_only") {
    detrend_climate = FALSE
    old_records_only = FALSE
    young_records_only = TRUE
  }
  
variables_dropped <- list() # this is to store the variables that were not conserdered for best model because the average % gap of the window >=5%
summary_data <- NULL # this will hold n_tree, n_cores and range of years
climate_interactions <- NULL # this is just to show case
best_model_summaries <- NULL
best_models_R_squared <- NULL
# ylim_p <- list() # this is to later readjust ylim in the GLS results plots, to standardize ylim across sites
all_Species_Year_to_keep <- NULL
species_removed_from_year_analysis <- NULL
## save every object names up until now to erase other stuff before runing each new site
data_to_keep <- c(ls(), "data_to_keep")


for(site in switch(solution_to_global_trend, "none" = sites[], c("ScottyCreek", "NewMexico", "SCBI")[])) {
  
  
  rm(list = ls()[!ls() %in% data_to_keep])
  cat(paste("running analysis for", site, "...\n" ))
  
  
  reference_date <- reference_dates[[site]]
  window_range <-  window_ranges[[site]]
  
  if(!detrend_climate & !old_records_only & !young_records_only) file.remove(list.files("results/residuals_by_tag_examples/", pattern = site, full.names = T))
  
  variables_dropped[[site]] <- list()
  
  for(what in switch(as.character(any(!is.na(all_Biol[[site]]$dbh))), "TRUE" = c("log_core_measurement", "log_core_measurement_dbh", "log_agb_inc_dbh", "log_BAI_dbh")[], "FALSE" = "log_core_measurement")) {
    
    Biol <- all_Biol[[site]]
    Clim <- droplevels(all_Clim[all_Clim$sites_sitename %in% sites.sitenames[site], ])

    ## create folders if don't exist ####
    dir.create(paste0("results/", what, "/", site), showWarnings = F, recursive = T)
    dir.create(paste0("results/with_CO2/", what, "/", site), showWarnings = F, recursive = T)
    dir.create(paste0("results/with_Year/", what, "/", site), showWarnings = F, recursive = T)
    
    dir.create(paste0("results/", solution_to_global_trend, "/", what, "/", site), showWarnings = F, recursive = T)
    
    ## remove all files so that we start clear #### 
    if(!detrend_climate & !old_records_only &!young_records_only) {
      file.remove(list.files(paste0("results/", what, "/", site), full.names = T))
    file.remove(list.files(paste0("results/with_CO2/", what, "/", site), full.names = T))
    file.remove(list.files(paste0("results/with_Year/", what, "/", site), full.names = T))
    } 
  
    file.remove(list.files(paste0("results/", ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, "/", site), full.names = T))
  
    
    ## if looking at older records only, remove data prior to date chosen ####
    if(old_records_only) Biol <- Biol[Biol$Year <= last_year_older_records, ]
    if(young_records_only) Biol <- Biol[Biol$Year > last_year_older_records, ]
    
    ## if we use dbh, remove any missing dbh ####
    if(grepl("dbh", what)) Biol <- droplevels(Biol[!is.na(Biol$dbh), ])
    
    ## if we use agb_inc, remove any missing agb_inc ####
    if(grepl("agb_inc", what)) Biol <- droplevels(Biol[!is.na(Biol$agb_inc), ])
    
    ## if we use BAI, remove any missing agb_inc ####
    if(grepl("BAI", what)) Biol <- droplevels(Biol[!is.na(Biol$BAI), ])

    ## remove any cores with fewer than 10 measurements ####
    Biol <- droplevels(Biol[Biol$coreID %in% names(which(tapply(Biol$core_measurement, Biol$coreID, length) >=10)), ])
    
    ## remove any species that has less than 7 different treeID (even if we use coreID as random effect) ####
    Biol <- droplevels(Biol[Biol$species_code %in% names(which(tapply(Biol$treeID, Biol$species_code, function(x) length(unique(x))) >= 7)), ])
    
    ## calculate residuals of spline measurement ~ year for each individual####
    Biol$residuals <- NA
    
    for( t in unique(Biol$coreID)) {
      x <- Biol[Biol$coreID %in% t, ]
      
      x$Y <- x[, switch(gsub("_dbh", "", what), 
                        log_core_measurement = "core_measurement", 
                        log_agb_inc = "agb_inc",
                        log_BAI = "BAI")]
      
      first_year_removed <- any(is.na(x$Y))
      x <- x[!is.na(x$Y),] #remove NA (only first year of measurement for agb_inc)
      
      test <- gam(Y~ s(Year), data = x) # see https://stats.stackexchange.com/questions/359568/choosing-k-in-mgcvs-gam
      
      
      if(!detrend_climate & !old_records_only & !young_records_only & rbinom(1, 1, 0.05)==1) {
        png(paste0('results/residuals_by_tag_examples/', paste(x$species_code[1], x$tree_status[1], t, sep = "_" ), "_", gsub("log_", "", what), "_Year_GAM", "_", site, '.png'),
            width = 8,
            height =8,
            units = "in",
            res = 300)


        par(mfrow = c(3,2))
        # plot(test)
        gam.check(test,pch=19,cex=.3)

        plot(Y~ Year, data = x, main = "Raw data", las = 1)
        
        plot(Y~ Year, data = x, main = "Raw data", las = 1)
        lines(test$fitted.values ~ x$Year)
        segments(x0 = x$Year, y0 = x$Y, x1 = x$Year, y1 = test$fitted.values)

        title(paste(x$species_code[1], x$tree_status[1], t, sep = " - " ), outer = T, line = -2)

        # save plot
        dev.off()
      }
      
      # save residuals
      x$residuals <- test$residuals 
      
      # remove Y
      x$Y <- NULL
      
      # save back into Biol
      if(gsub("_dbh", "", what) %in% c("log_agb_inc", "log_BAI") & first_year_removed) {
        Biol[Biol$coreID %in% t, ]$residuals <- c(NA, x$residuals) 
      } else {
        Biol[Biol$coreID %in% t, ] <- x
      }
      
      
    }
    
    
    ## run slidingwin on residuals to find best time window with quad for each variable ####
    if(nlevels(Biol$species_code) > 3) {
      baseline = "lmer(residuals ~ 1 + (1 | species_code) + (1 | coreID), data = Biol[!is.na(Biol$residuals), ])"
      
    } else 
    {
      baseline = "lmer(residuals ~ 1 + (1 | coreID), data = Biol[!is.na(Biol$residuals), ])"
    }
    
    
    results <- slidingwin( baseline = eval(parse(text = baseline)),
                           xvar =  list(
                             pet = Clim[,paste0("pet", ifelse(detrend_climate, "_detrended", ""))], 
                             tmn = Clim[,paste0("tmn", ifelse(detrend_climate, "_detrended", ""))], 
                             tmp = Clim[,paste0("tmp", ifelse(detrend_climate, "_detrended", ""))], 
                             tmx = Clim[,paste0("tmx", ifelse(detrend_climate, "_detrended", ""))],
                             pre = Clim[,paste0("pre", ifelse(detrend_climate, "_detrended", ""))], 
                             wet = Clim[,paste0("wet", ifelse(detrend_climate, "_detrended", ""))],
                             SF = Clim$SF # not detrended because nodon't ru at Niabora
                           )
                           [switch(site, "CedarBreaks" = c( "tmn", "tmp", "tmx", "pre"), "NewMexico" = c( "tmn", "tmp", "tmx", "pre"), Niobara = c("pet", "tmn", "tmp", "tmx", "pre", "wet", "SF"), c("pet", "tmn", "tmp", "tmx", "pre", "wet"))],
                           type = "absolute", 
                           range = window_range,
                           stat = c("mean"),
                           func = "quad", # c("lin","quad")
                           refday = reference_date,
                           cinterval = "month",
                           cdate = Clim$Date, bdate = Biol[!is.na(Biol$residuals), ]$Date,
                           cmissing = "method1") 
    
    
    # # change best windows to the 95 CI (see ssue #95) ####
    # names(results$combos) <- c("response", "climate", "type", "stat", "func", "original_DeltaAICc", "original_WindowOpen", "original_WindowClose")
    # mytry <- NULL
    # for(i in 1:nrow(results$combos)) {
    #   mytry <- rbind(mytry, cbind(results$combos[i, ], get_95_CI_window_showed_in_plot(dataset = results[[i]]$Dataset)))
    # }
    # if(!all(order(mytry$original_DeltaAICc) == order(mytry$deltaAICc))) stop("not same order of AICc")
#   }
# }
# }
    
    
# Check if best windows for each variable meet the maximum gap requirement (average of month-variale combination <= 5% of the time period) ####
    adjusted_clim_gap <- adjusted_clim_gaps[[site]]
    variables_dropped_site <- NULL
    
    for(i in 1:nrow(results$combos)) {
      
      months_to_avg <- results$combos$WindowOpen[i]:results$combos$WindowClose[i]
      months_to_avg <- reference_date[2] - months_to_avg
      # months_to_avg <- months_to_avg[months_to_avg!= 0 ]
      months_to_avg <- ifelse(months_to_avg<=0, rev(1:12)[abs(months_to_avg)+1], months_to_avg)
      
      idx_v <- adjusted_clim_gap$start_climvar.class %in% results$combos$climate[i]
      
      if(any(idx_v) & mean(adjusted_clim_gap[idx_v, ][match(months_to_avg, adjusted_clim_gap$month[idx_v]),]$percent_misssing) > clim_gap_threshold ) {
        variables_dropped_site <- c(variables_dropped_site, as.character(results$combos$climate[i]))
      } 
    }
    variables_dropped[[site]][[what]] <- variables_dropped_site
    
    ### find best window for each variable group ignoring the variables that need to be dropped because of gap filling issues (best way to not get confused later with the ordering of te slindingwin order) ####
    clim_var_group_site <- lapply(clim_var_group, function(x) x[!x %in% variables_dropped[[site]][[what]]])
  
    best_results_combos <- do.call(rbind, lapply(clim_var_group_site, function(X) {
      x <- results$combos[results$combos$climate %in% X,]
      data.frame(model_ID = as.numeric(rownames(x)), x, stringsAsFactors = F)[which.min(x$DeltaAICc),]
    }))
    ### plot climwin results ####
    for(i in best_results_combos$model_ID) {
      png(paste0('results/', ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, '/', site, '/climwin_', paste((data.frame(lapply(results$combos[i, c(2, 5, 7, 8)], as.character), stringsAsFactors=FALSE)), collapse = "_"), "_", site, '.png'),
          width = 10,
          height =8,
          units = "in",
          res = 300)
      # plot the results
      plotall(dataset = results[[i]]$Dataset, 
              bestmodel = results[[i]]$BestModel,
              bestmodeldata = results[[i]]$BestModelData,
              title=paste((data.frame(lapply(results$combos[i,], as.character), stringsAsFactors=FALSE)), collapse = "_"), arrow = T)
      
      # dev.off()
      dev.off()
      
    }
    
    # # if PET does not come out as top variable in both T and CLD group, drop it (commeted out) ####
    # if(any(duplicated(best_results_combos))) { best_results_combos <- best_results_combos[!duplicated(best_results_combos), ]# remove one pet as it came out best in both T and cld group
    # } else {
    #   best_results_combos <- best_results_combos[!best_results_combos$climate %in% "pet", ] # remove pet as it was bitten in either T or CLD group
    # }
    # best_results_combos <- best_results_combos[order(best_results_combos$DeltaAICc),]
    # 
    ### Save the signal into Biol ####
    for(i in best_results_combos$model_ID) {
      print(paste("adding climate data to Biol for model", i, "in", what))
     
      
      # if(any(grepl("I\\(climate\\^2\\)", names( results[[i]]$BestModelData)))) {
      #   columns_to_add <- results[[i]]$BestModelData[, c("climate", "I(climate^2)")]
      #   names(columns_to_add) <- paste0(results$combos[i,]$climate, c("", "^2"))
      # } else {
        columns_to_add <- results[[i]]$BestModelData[c("climate")]
        names(columns_to_add) <- results$combos[i,]$climate
      # } if(any(grepl("I\\(climate\\^2\\)", names( results[[i]]$BestModelData)))) 
      
      to_add_to_Biol <-  cbind(Biol[!is.na(Biol$residuals), ], columns_to_add)
      Biol[, names(columns_to_add)] <- to_add_to_Biol[match(rownames(Biol), rownames(to_add_to_Biol)), names(columns_to_add)]
    }
    
    ## Output Biol and best_results_combos to use in different analysis ####
    dir.create(paste0("processed_data/core_data_with_best_climate_signal/", what, "/"), recursive = T, showWarnings = F)
    write.csv(Biol, file = paste0("processed_data/core_data_with_best_climate_signal/", what, "/", site, ifelse(solution_to_global_trend %in% "none", "", paste0("_", solution_to_global_trend)), ".csv"), row.names = F)
  
    
   
    # now do a species by species gls using log of raw measurements, spline on dbh and year (WITH AND WITHOUT CO2 or Year) ####
    
    ## first check if we are allowed to try Year or CO2 based on this criteria ####
    ### we limited the analysis of each species to years for which the dbh range was represented by at least 3 individuals and spanned at least half the total DBH range of that species
    
    ## look at range of dbh and Year for each species
    dbh_range_by_species <- aggregate(dbh ~ species_code, data = Biol, FUN = function(x) max(x) - min(x))
    year_range_by_species <- aggregate(Year ~ species_code, data = Biol, FUN = function(x) max(x) - min(x))

    ## look at range of dbh for each year and each species
    dbh_range_by_Year_and_species <- aggregate(dbh ~ Year + species_code, data = Biol, FUN = function(x) max(x) - min(x))
    
    ## for each species, look at if the range of dbh in a a year is at least half the total range of dbh for that species
    dbh_range_by_Year_and_species$sp_dbh_range <- dbh_range_by_species$dbh[match(dbh_range_by_Year_and_species$species_code, dbh_range_by_species$species_code)]
    
    dbh_range_by_Year_and_species$enough_dbh_range <- dbh_range_by_Year_and_species$dbh >= dbh_range_by_Year_and_species$sp_dbh_range/2.5 # 2 for hal, 3 for 1/3rd) 2.5 for 40%
  
    
    ## keep a list of species x year combo to keep in analysis.
    Species_Year_to_keep <- dbh_range_by_Year_and_species[dbh_range_by_Year_and_species$enough_dbh_range, c("species_code", "Year")]
    
    
    ## for each species, look at if the total n years meetin dbh criteria above is enough 
    year_range_by_species$period_range_limit <- year_range_by_species$Year/3*2 # 2 third of the total 
    
    if(solution_to_global_trend %in% "none" & any(table(Species_Year_to_keep$species_code)<year_range_by_species$period_range_limit)) species_removed_from_year_analysis <- unique(rbind(species_removed_from_year_analysis, data.frame(site, species_code = names(which(table(Species_Year_to_keep$species_code)< year_range_by_species$period_range_limit )))))
    
    ## remove any species that have less than the criteria n years
    Species_Year_to_keep <- Species_Year_to_keep[Species_Year_to_keep$species_code %in% names(which(table(Species_Year_to_keep$species_code)>=year_range_by_species$period_range_limit)), ]
    
    ## keep a record of that
    
    if(solution_to_global_trend %in% "none") all_Species_Year_to_keep <- unique(rbind(all_Species_Year_to_keep,                             data.frame(site,                                                 Species_Year_to_keep)))
    
    ## then run GLS ####
   for(with_Year_or_CO2 in switch(as.character(solution_to_global_trend == "none" & grepl("dbh", what) & nrow(Species_Year_to_keep)>0), "TRUE" = c("", "Year", "CO2")[-3], "FALSE" = "")) {
      
      ## look at collinearity between climate variables ( and CO2 and Year and dbh when relevant) and remove any variable with vif > 10 ####
    if(grepl("dbh", what) & !with_Year_or_CO2 %in% "") X <- Biol[, c(as.character(best_results_combos$climate), with_Year_or_CO2, "dbh")]
    if(grepl("dbh", what) & with_Year_or_CO2 %in% "") X <- Biol[, c(as.character(best_results_combos$climate), "dbh")]
    if(!grepl("dbh", what) & !with_Year_or_CO2 %in% "") X <- Biol[, c(as.character(best_results_combos$climate), with_Year_or_CO2)]
    if(!grepl("dbh", what) & with_Year_or_CO2 %in% "") X <- Biol[, c(as.character(best_results_combos$climate))]
    
    X <- X[!duplicated(X),]
    
    
    usdm::vif(X)
    (vif_res <-  usdm::vifstep(X, th = 3))
    sink(paste0("results/", ifelse(with_Year_or_CO2 %in% "", "", paste0("with_", with_Year_or_CO2, "/")), ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, "/", site, "/VIF.txt"))
    print(vif_res)
    sink()
    variables_to_keep <- as.character(vif_res@results$Variables)
    
    
    ## create the gls formula ####
    
    full_model_formula <- switch(gsub("_dbh", "", what), 
                                 "log_core_measurement" = paste("log_core_measurement ~", paste(variables_to_keep, "+", paste0("I(", variables_to_keep, "^2)"), collapse = " + ")),
                                 "log_agb_inc" = paste("log_agb_inc ~", paste(variables_to_keep, "+", paste0("I(", variables_to_keep, "^2)"), collapse = " + ")),
                                 "log_BAI" = paste("log_BAI ~", paste(variables_to_keep, "+", paste0("I(", variables_to_keep, "^2)"), collapse = " + ") ))
    
    # prepare subset for dredge
    sexpr <-parse(text = paste("!(", paste0("(`I(", variables_to_keep, "^2)` & !", variables_to_keep, ")", collapse = "|"), ")"))
    
    # remove second order term for CO2 and Year
    full_model_formula <- gsub("I\\(CO2\\^2\\) \\+ |I\\(Year\\^2\\) \\+ ", "", full_model_formula)
    sexpr <- parse(text =gsub("\\| \\(`I\\(CO2\\^2\\)` & \\!CO2\\)|\\| \\(`I\\(Year\\^2\\)` & \\!Year\\)", "", sexpr))
   
    
    ## identify what variables we should keep for each species, looking at the sum of AIC weights ####
    
    if( with_Year_or_CO2 %in% "") species_to_keep <- unique(Biol$species_code) else species_to_keep <- unique(Species_Year_to_keep$species_code)
    
    
    sum_of_weights_for_each_term_by_sp <- NULL
    rm(list = ls()[grepl("_best_model_int_", ls())])
    for(sp in species_to_keep) {
      start_time <- Sys.time()
      x <- Biol[Biol$species_code %in% sp,]
      
      # keep only years with enough data if we look at CO2 or Year
      # if(!with_Year_or_CO2 %in% "") x <- x[x$Year %in% Species_Year_to_keep$Year[Species_Year_to_keep$species_code %in% sp],]
      
      cat("Running GLS and dredging for species", sp , "and its", length(unique(x$coreID)), "trees...\n")
      
      # log transform the response variables
      x$log_core_measurement <- log(x$core_measurement) 
      x$log_agb_inc <-  log(x$agb_inc)
      x$log_BAI <-  log(x$BAI)
      
      # replace the undefined log values by the log of half the smallest value
    x$log_core_measurement[x$core_measurement %in% 0] <- log(min( x$core_measurement[x$core_measurement >0])/2)
    x$log_agb_inc[x$agb_inc %in% 0] <- log(min( x$agb_inc[x$agb_inc >0])/2)
    x$log_BAI[x$BAI %in% 0] <- log(min( x$BAI[x$BAI >0])/2)
      # x <- x[, c("dbh", "Year", "tag", what, variables_to_keep)]
      
      x <- x[, c("Year", "treeID", "coreID", "dbh", "CO2", gsub("_dbh", "", what), variables_to_keep)]
      
      x <- droplevels(x[!is.na(x[, gsub("_dbh", "", what)]), ])
      
      fm1 <- lme((eval(parse(text = full_model_formula))), random = ~1|coreID, correlation = corCAR1(form=~Year|coreID), data = x, na.action = "na.fail", method = "ML") 
      
      clust <- makeCluster(getOption("cl.cores", 2), type = "SOCK")
      clusterEvalQ(clust, library(splines))
      clusterEvalQ(clust, library(nlme))
      clusterExport(clust, list  = c("x", "fm1", "sexpr"))
      dd <- pdredge(fm1, trace = 2, cluster = clust, subset = sexpr)
      stopCluster(clust)
      dd$cw <- cumsum(dd$weight)
      
      # add a column in dd if any u-shaped and so model not considered in best model
      columns_to_check_for_ushape <- grep(paste(c(climate_variables, "SF"), collapse = "|"), names(dd), value = T)
      dd$any_u_shaped <-   apply(dd[, columns_to_check_for_ushape], 1, function(x) any(tapply(as.numeric(sign(x)), rep(1:(length(x)/2), each=2), paste, collapse = "") %in% c("-11")))
       
      # save dd
      write.csv(dd, file = paste0("results/", ifelse(with_Year_or_CO2 %in% "", "", paste0("with_", with_Year_or_CO2, "/")), ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, "/", site, "/", sp, "_model_comparisons.csv"), row.names = F)
      
      # get sum of weights
      # sum_of_weights_for_each_term <- dd[, grepl(paste(c(variables_to_keep, "dbh", "Year"), collapse = "|"), names(dd))]
      sum_of_weights_for_each_term <- dd[, grepl(paste(c(variables_to_keep,  "dbh", "Year", "coreID"), collapse = "|"), names(dd))]
      sum_of_weights_for_each_term <- apply(sum_of_weights_for_each_term, 2, function(x) sum(dd$weight[!is.na(x)]))
      sum_of_weights_for_each_term
      sum_of_weights_for_each_term_by_sp <- rbind(sum_of_weights_for_each_term_by_sp,
                                                  c(sum_of_weights_for_each_term))
      
      
      
      # update the results of the best model to REML = T
      best_model <- update(get.models(dd, which(dd$any_u_shaped==0)[1])[[1]], method = "REML") # take first model without a u-shape in climate variable
      # get the results of the model that includes the variables that have sum of weight > 0.9
      # if(sum(sum_of_weights_for_each_term[!grepl("(coreID|Year)", names(sum_of_weights_for_each_term))] > 0.9) == 0 ) {
      #   best_model <- update(fm1, fixe = ~1, method = "REML") #intercept only model if none of the variables have 0.9 sum of weitgh
      # } else {
      #   best_model <- update(fm1, fixe = paste( "~", paste(names(sum_of_weights_for_each_term)[sum_of_weights_for_each_term > 0.9], collapse = " + ")), method = "REML")
      # }
      
      
      
      # save results for individual species
      assign(paste0(sp, "_dd"), dd)
      assign(paste0(sp, "_best_model"), best_model)
      
      # save into best_model_summaries
      best_model_summaries <- rbind(
        best_model_summaries,
        data.frame(with_Year_or_CO2,
                   what,
                   site,
                   sp,
                   parameter = rownames(summary(best_model)$tTable),
                   summary(best_model)$tTable)
      )
      best_models_R_squared <- rbind(best_models_R_squared,
                                     data.frame(with_Year_or_CO2,
                                                what,
                                                site,
                                                sp,
                                                model =as.character(best_model$call[2]),
                                                r.squaredGLMM(best_model)))
      
      # keeping track of timing
      end_time <- Sys.time()
      (ellapsed_time <- difftime(end_time, start_time, units = "mins"))
  
      
      # try interaction between DBH and climate if we are not looking at Year or CO2, but if there is the same criteria met than for trying CO2 and Yeart ####
   
      if(grepl("dbh", what) & with_Year_or_CO2 %in% "" & "dbh" %in% variables_to_keep & !detrend_climate & !old_records_only & !young_records_only & sp %in% Species_Year_to_keep$species_code) {
        
        # x <- x[x$Year %in% Species_Year_to_keep$Year[Species_Year_to_keep$species_code %in% sp],]
        
       
        for(g in names(clim_var_group)) {
          v_int <- names(which(sapply(clim_var_group[[g]], grepl, as.character(best_model$call[2]))))
        
          rm(list = paste0(sp, "_best_model_int_", g))
          
          if(length(v_int) > 0 ) {
            bm_clim_int <- update(best_model, fixe = as.formula(paste0("~ . - I(dbh^2) + dbh:", v_int)))
            climate_interactions <- 
              rbind(climate_interactions,                                         data.frame(
                what,
                site,
                species = sp,
                climate_group = g,
                model_formula = as.character(bm_clim_int$call[2]),
                variable = v_int,
                interaction_coef = fixef(bm_clim_int)[grepl(":", names(fixef(bm_clim_int)))],
                p_value = summary(bm_clim_int)$tTable[grepl(":", rownames(summary(bm_clim_int)$tTable)), "p-value"]
              )
              )
            
            if(summary(bm_clim_int)$tTable[grepl(":", rownames(summary(bm_clim_int)$tTable)), "p-value"] < 0.05)             assign(paste0(sp, "_best_model_int_", g), bm_clim_int) # keep model if intereaction is significant

          }
          
          
          } #  for(g in names(clim_var_group)
       
        
        
        
      } # do dbh x climate if in good scenario
      
      # remove x
      rm(x)
    } #sp in species_to_keep
    
    
    rownames(sum_of_weights_for_each_term_by_sp) <- species_to_keep
    
    sum_of_weights_for_each_term_by_sp
    
    # save the plot
    png(paste0('results/', ifelse(with_Year_or_CO2 %in% "", "", paste0("with_", with_Year_or_CO2, "/")), ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, "/", site, "/GLS_Sum_of_AICweights_", site, '.png'),
        width = 10,
        height =8,
        units = "in",
        res = 300)
    print(lattice::levelplot(t(sum_of_weights_for_each_term_by_sp), 
                       scales=list(x=list(rot=45)), 
                       xlab = "parameter", 
                       ylab = "species",
                       legend = list(top = list(fun = grid::textGrob("Sum of Weights", y=0, x=1.09)))))
    
    Sys.sleep(time = 1) # to give time for plot to show up
    #dev.off()
    dev.off()
    
    
    
    ## Plot response curves for each variable, with one curve per species ####
    # first remove any object starting by p_
    rm(list = ls()[grepl("^p_", ls())])
    #Create a custom color scale
    ylim_p <- list()
    ylim_p_int <- list()
    for(v in c( variables_to_keep)) { 
      print(v)
      
      ## predictions
      pt <- NULL
      pt_int <- NULL
      for(sp in rownames(sum_of_weights_for_each_term_by_sp)) {
        best_model <- get(paste0(sp, "_best_model"))
        
        g <- names(which(sapply(clim_var_group, function(x) v %in% x)))
        
        if(exists(paste0(sp, "_best_model_int_", g))) {
          dbh_clim_int = T
          best_model_int <- get(paste0(sp, "_best_model_int_", g))
        } else {
          dbh_clim_int = F
        }

        
        x <- Biol[Biol$species_code %in% sp,]
        
        # keep only years with enough data if we look at CO2 or Year
        # if(!with_Year_or_CO2 %in% "") x <- x[x$Year %in% Species_Year_to_keep$Year[Species_Year_to_keep$species_code %in% sp],]
        
        varying_x <- data.frame(varying_x = seq(min(x[, v], na.rm = T), max(x[, v], na.rm = T), length.out = 100)) ; colnames(varying_x) <- v
        # constant_variables <- c("dbh", variables_to_keep)[!c("dbh", variables_to_keep) %in% v]
        constant_variables <- c(variables_to_keep)[!c(variables_to_keep) %in% v]
        
        newd <- cbind(eval(parse(text = paste0("data.frame(", paste0(constant_variables, " = median(x$", constant_variables, ", na.rm = T)", collapse = ", "), ")"))), varying_x)  # newd <- cbind(eval(parse(text = paste0("data.frame(", paste0(constant_variables, " = median(Biol$", constant_variables, ", na.rm = T)", collapse = ", "), ",  Year = median(Biol$Year, na.rm = T), coreID = factor(Biol[Biol$species_code %in% sp,]$coreID[1]))"))), varying_x)
        
        
      
        # if(v %in% names(best_model$var.summary)) {
        # pt <- rbind(pt, data.frame(newd, variable = v, species = sp, varying_x = newd[, v], do.call(rbind, lapply(1:nrow(newd), function(i) data.frame(predict(best_model, newd[i,], type = "link", level = 0, se.fit = T)))), draw = any(grepl(v,names(best_model$coefficients$fixed)))))
       
        
        pt <- rbind(pt, 
                    data.frame(newd, 
                               variable = v, 
                               species = sp, 
                               varying_x = newd[, v], 
                               data.frame(predict(best_model, newd, type = "link", level = 0, se.fit = T)), 
                               sigma = sigma(best_model), # get residual Standard error to correct prediction of mean when back-transforming
                               draw = any(grepl(v,names(best_model$coefficients$fixed))),
                               sig = switch(as.character(sum(summary(best_model)$tTable[grep(v, rownames(summary(best_model)$tTable)), "p-value"] < .05)), "2" = "solid", "1" = "twodash", "0" = "dotted")))
        
        
        if(dbh_clim_int) {
          constant_variables <- constant_variables[!constant_variables%in%"dbh"]
          newd_int <- cbind(eval(parse(text = paste0("data.frame(", paste0(constant_variables, " = median(x$", constant_variables, ", na.rm = T)", collapse = ", "), ", dbh = range(x$dbh))"))), varying_x[rep(1:nrow(varying_x), each = 2),, drop = F]) 
      
        
        pt_int <- rbind(pt_int, 
                        data.frame(newd_int, 
                                   variable = v, 
                                   species = sp, 
                                   varying_x = newd_int[, v], 
                                   data.frame(predict(best_model_int, newd_int, type = "link", level = 0, se.fit = T)), 
                                   sigma = sigma(best_model_int), # get residual Standard error to correct prediction of mean when back-transforming
                                   draw = any(grepl(v, names(best_model_int$coefficients$fixed))),
                                   # sig = switch(as.character(sum(summary(best_model_int)$tTable[grep(v, rownames(summary(best_model_int)$tTable)), "p-value"] < .05)), "2" = "solid", "1" = "twodash", "0" = "dotted"),
                                   DBH = factor(c("min", "max")[as.numeric(as.factor(newd_int$dbh))])
                                   ))
      }
        # }
        
      } # ignore warnings
      
      # if(!is.null(pt)) {
      pt$species_code <- pt$species
      pt$species <- factor(pt$species, levels = rownames(sum_of_weights_for_each_term_by_sp))
      pt$species <- factor(paste0(pt$species, " (",tapply(Biol$coreID,  Biol$species_code, function(x) length(unique(x)))[as.character(pt$species)], ")"))
      pt$expfit <- exp(pt$fit + .5*pt$sigma^2) # exp(pt$fit) would give epected median.
      pt$lwr <- exp(pt$fit + .5*pt$sigma^2 - 1.96 * pt$se.fit)# exp(pt$fit - 1.96 * pt$se.fit)
      pt$upr <- exp(pt$fit + .5*pt$sigma^2 + 1.96 * pt$se.fit)
     
      pt$sig <- as.character(pt$sig)
      
      species_colors <- all_species_colors[[site]]
      names(species_colors) <- pt$species[match(names(species_colors), pt$species_code)]
      species_colors <- species_colors[!is.na(names(species_colors))]
      
      
      time_window <- reference_date[2] - as.numeric(best_results_combos[best_results_combos$climate %in% v, c("WindowOpen", "WindowClose")])
      time_window_prev <- time_window <= 0
      time_window <- ifelse(time_window_prev, rev(1:12)[abs(time_window) +1], time_window)
      
      time_window_text <- paste(paste0(ifelse(time_window_prev, "p.", "c."),month.abb[time_window]), collapse = "-") #  paste0("\nfrom ", paste(paste0(ifelse(time_window_prev, "prev. ", "curr. "), month.abb[time_window]), collapse = "\nto "))
      
      
      p <- ggplot(data = pt[pt$draw,], aes(x = varying_x, y = expfit))
     
      # add gray rectangle when relevant
      if(!v %in% c("dbh", "Year")) p <- p + geom_rect(xmin = mean(Biol[, v], na.rm = T) - sd(Biol[, v], na.rm = T), ymin = min(pt$lwr), xmax = mean(Biol[, v], na.rm = T) + sd(Biol[, v], na.rm = T), ymax = max(pt$upr), fill = "grey" , alpha=0.01) + geom_vline(xintercept = mean(Biol[, v], na.rm = T), col = "grey")
      

      p <- p + geom_line(aes(col = species, linetype = sig)) +
        # scale_x_continuous(trans= ifelse(v %in% "dbh", 'log','identity')) +
        labs(#title = paste0(v, ifelse(v %in% best_results_combos$climate, time_window_text, "")),
             x = paste(v, ifelse(v %in% best_results_combos$climate, time_window_text, ""), variables_units[[v]]),
             y = "") + #"core measurements") +
        geom_ribbon(aes(ymin=lwr, ymax=upr, bg = species), alpha=0.25) +
        scale_color_manual(values = species_colors[unique(pt$species[pt$draw])]) + scale_fill_manual(values = species_colors[unique(pt$species[pt$draw])]) +
        # scale_colour_hue(drop = F) + scale_fill_hue(drop = F) + 
        theme_classic() +
        theme(text = element_text(size = 10)) + 
        scale_linetype_manual(values = c("solid" = 1, "dashed" = 2, "dotted"= 3 , "dotdash" = 4, "longdash" = 5, "twodash" = 6))+
        guides(linetype = FALSE)
      
      if(any(pt$draw) | v %in% "Year") {
        assign(paste0("p_", v), p +
               theme(legend.position="none")) 
      }
      
      # save the ylimits ####
      ylim_p[[v]] <- range(pt$lwr, pt$upr)
      # if(!with_CO2) ylim_p[[what]][[v]][[site]] <- range(pt$lwr, pt$upr)
      
      
      if(!is.null(pt_int)) {
        pt_int$species_code <- pt_int$species
        pt_int$species <- factor(pt_int$species, levels = rownames(sum_of_weights_for_each_term_by_sp))
        pt_int$species <- factor(paste0(pt_int$species, " (",tapply(Biol$coreID,  Biol$species_code, function(x) length(unique(x)))[as.character(pt_int$species)], ")"))
        pt_int$expfit <- exp(pt_int$fit + .5*pt_int$sigma^2) # exp(pt_int$fit) would give epected median.
        pt_int$lwr <- exp(pt_int$fit + .5*pt_int$sigma^2 - 1.96 * pt_int$se.fit)# exp(pt_int$fit - 1.96 * pt_int$se.fit)
        pt_int$upr <- exp(pt_int$fit + .5*pt_int$sigma^2 + 1.96 * pt_int$se.fit)
      
        
        p_int <- ggplot(data = pt_int[pt_int$draw,], aes(x = varying_x, y = expfit, group = dbh))
        
        # add gray rectangle when relevant
        if(!v %in% c("dbh", "Year")) p_int <- p_int + geom_rect(xmin = mean(Biol[, v], na.rm = T) - sd(Biol[, v], na.rm = T), ymin = min(pt$lwr), xmax = mean(Biol[, v], na.rm = T) + sd(Biol[, v], na.rm = T), ymax = max(pt$upr), fill = "grey" , alpha=0.01) + geom_vline(xintercept = mean(Biol[, v], na.rm = T), col = "grey")
        
        
        p_int <- p_int + geom_line(aes(col = species, linetype = DBH)) +
          # scale_x_continuous(trans= ifelse(v %in% "dbh", 'log','identity')) +
          labs(#title = paste0(v, ifelse(v %in% best_results_combos$climate, time_window_text, "")),
            x = paste(v, ifelse(v %in% best_results_combos$climate, time_window_text, ""), variables_units[[v]]),
            y = "") + #"core measurements") +
          geom_ribbon(aes(ymin=lwr, ymax=upr, bg = species), alpha=0.25) +
          scale_color_manual(values = species_colors[unique(pt$species[pt$draw])]) + scale_fill_manual(values = species_colors[unique(pt$species[pt$draw])]) +
          # scale_colour_hue(drop = F) + scale_fill_hue(drop = F) + 
          theme_classic() +
          theme(text = element_text(size = 10))
        
        if(any(pt_int$draw) | v %in% "Year") {
          assign(paste0("p_int_", v), p_int) 
        }
        
        # save the ylimits ####
        ylim_p_int[[v]] <- range(pt_int$lwr, pt_int$upr)
        # if(!with_CO2) ylim_p[[what]][[v]][[site]] <- range(pt$lwr, pt$upr)
      }
     #  p$data <- pt_int
     #  p <- p + aes(group = factor(p$data$dbh))
     # p$layers[[3]]$aes_params$linetype <- as.character(p$data$sig)
     # p$layers[[3]]$aes_params$size <- p$data$linesize*0.5


    }
    
    g_legend <- function()
    {
      a.gplot <- ggplot(data = pt, aes(x = varying_x, y = expfit)) + geom_line(aes(group = species, col = species)) +  geom_ribbon(aes(ymin=lwr, ymax=upr, bg = species), alpha=0.25) +
        scale_color_manual(values = species_colors) + scale_fill_manual(values = species_colors)
        tmp <- ggplot_gtable(ggplot_build(a.gplot))
        leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
        legend <- tmp$grobs[[leg]]
        return(legend)
    }
    
    # existing_plots <- paste0("p_",  c("dbh", variables_to_keep))
    existing_plots <- ls()[grepl("^p_[^in]", ls())]
    existing_plots_int <- ls()[grepl("^p_int_", ls())]
    
    grid.arrange(do.call(arrangeGrob, c(lapply(existing_plots, function(x) {
      x <- get(x)
      x$layers[[1]]$aes_params$ymin <- range(ylim_p)[1]
      x$layers[[1]]$aes_params$ymax <- range(ylim_p)[2]
      x + ylim(range(ylim_p))
      }), ncol = ifelse(length(existing_plots) %in% 4, 2, length(existing_plots)))),
                 g_legend(),
                 nrow = 1,
                 widths = c(10, 1))
    
     grid::grid.text(switch (gsub("_dbh", "", what), log_core_measurement = "core measurement (mm)",
                            log_agb_inc = "AGB increment (Mg C)", log_BAI = "BAI (cm2)"), x = unit(0.01, "npc"), y = unit(.51, "npc"), rot = 90, gp = grid::gpar(fontsize = 10))
    
    
    
    # save png plot ####
    png(paste0('results/', ifelse(with_Year_or_CO2 %in% "", "", paste0("with_", with_Year_or_CO2, "/")), ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, "/", site, "/GLS_ALL_variables_responses_", site, '.png'),
        width = 10,
        height =8,
        units = "in",
        res = 300)
    
    grid.arrange(do.call(arrangeGrob, c(lapply(existing_plots, function(x) {
      x <- get(x)
      x$layers[[1]]$aes_params$ymin <- range(ylim_p)[1]
      x$layers[[1]]$aes_params$ymax <- range(ylim_p)[2]
      x + ylim(range(ylim_p))
    }), ncol = ifelse(length(existing_plots) %in% 4, 2, length(existing_plots)))),
    g_legend(),
    nrow = 1,
    widths = c(10, 1))
    
    grid::grid.text(switch (gsub("_dbh", "", what), log_core_measurement = "core measurement (mm)",
                            log_agb_inc = "AGB increment (Mg C)",
                            log_BAI = "BAI (cm2)"), x = unit(0.01, "npc"), y = unit(.51, "npc"), rot = 90, gp = grid::gpar(fontsize = 10))
    dev.off()
     
    # save interaction png plot ####
    if(length(existing_plots_int)>0) {
    png(paste0('results/', what, "/", site, "/GLS_ALL_variables_responses_", site, '_dbh_interaction.png'),
        width = 10,
        height =8,
        units = "in",
        res = 300)
    
      grid.arrange(do.call(arrangeGrob, c(lapply(existing_plots_int, function(x) {
        x <- get(x)
        x$layers[[1]]$aes_params$ymin <- range(ylim_p_int)[1]
        x$layers[[1]]$aes_params$ymax <- range(ylim_p_int)[2]
        x + ylim(range(ylim_p_int))
      }), ncol = ifelse(length(existing_plots_int) %in% 4, 2, length(existing_plots_int)))),
      nrow = 1,
      widths = c(10, 1))
    
    grid::grid.text(switch (gsub("_dbh", "", what), log_core_measurement = "core measurement (mm)",
                            log_agb_inc = "AGB increment (Mg C)",
                            log_BAI = "BAI (cm2)"), x = unit(0.01, "npc"), y = unit(.51, "npc"), rot = 90, gp = grid::gpar(fontsize = 10))
    dev.off()
    
    }
    # save plots at this point to later fetch them  ####
    save(list = grep("^p_|pt|clim_var_group$|ylim_p|species_colors", ls(), value = T), file = paste0('results/', ifelse(with_Year_or_CO2 %in% "", "", paste0("with_", with_Year_or_CO2, "/")), ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, "/", site, "/env.RData"))
    
   
    # save environment ####
    save.image(file = paste0("results/",  ifelse(with_Year_or_CO2 %in% "", "", paste0("with_", with_Year_or_CO2, "/")), ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), what, "/", site, "_all_env.RData"))
    
    } # for(with_Year_or_CO2 in ...) 
     
   # summarize data used in this analysis ####
    summary_data <- rbind(summary_data, data.frame(site = site, what = what,  
                          Biol %>% group_by(species_code) %>%
                            summarize(
                                      n_trees = n_distinct(treeID),
                                      n_cores = n_distinct(coreID),
                                      min_DBH_cores_sampled = min(dbh_at_coring),
                                      max_DBH_cores_sampled = max(dbh_at_coring),
                                      min_DBH_cores_reconstructed = round(min(dbh), 1),
                                      max_DBH_cores_reconstructed = round(max(dbh), 1),
                                      start_year = min(Year),
                                      end_year = max(Year)                                    )))

    
  } # for (what in ...)
  
  
  # save environment ####
  # if(!detrend_climate & !old_records_only & !young_records_only) save.image(file = paste0("results/", site, "_all_env.RData"))
} # for sites in ..

# save summary_data ####
if(solution_to_global_trend %in% "none") write.csv(summary_data, "results/summary_cores_analyzed.csv", row.names = F)

# save best_models summaries ####
write.csv(best_model_summaries, file = paste0("results/", ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), "best_model_summaries.csv"), row.names = F)
write.csv(best_models_R_squared, file = paste0("results/", ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), "best_models_R_squared.csv"), row.names = F)
# save and summarize climate_interactions ####
if(!detrend_climate & !old_records_only & !young_records_only) {
  write.csv(climate_interactions, file = "results/climate_interactions_coeficients.csv", row.names = F)
  
  n_sig <- aggregate(p_value ~ what + site + climate_group, data = climate_interactions, FUN = function(x) sum(x<0.05))
  n_species <- aggregate(species ~ what + site + climate_group, data = climate_interactions, FUN = function(x) length(unique(x)))
  freq_sig <- aggregate(p_value ~ what + site + climate_group, data = climate_interactions, FUN = function(x) round(sum(x<0.05)*100/length(x), 2))
  
  names(n_sig) <- gsub("p_value", "n_sig", names(n_sig))
  names(n_species) <- gsub("species", "n_species", names(n_species))
  names(freq_sig) <- gsub("p_value", "freq_sig", names(freq_sig))
  
  
climate_interactions_summary <- left_join(left_join(n_sig, n_species), freq_sig)

write.csv(climate_interactions_summary, file = "results/climate_interactions_summary.csv", row.names = F)
} # if(!detrend_climate)
# save variables_dropped ####
variables_dropped

sink(paste0("results/", ifelse(solution_to_global_trend %in% "none", "", paste0(solution_to_global_trend, "/")), "variables_dropped_at_each_site_due_to_gap.txt"))
variables_dropped
sink()


# # see VIF situations for all sites ####
# VIF_files <- list.files("results", pattern = "VIF.txt", recursive = T, full.names = T)
# VIF_files <- VIF_files[!grepl("CO2", VIF_files)]
# 
# # check if any variables were kicked because of linearity issues
# for(f in VIF_files) {
#   x <- readLines(f)[1]
#   if(substr(x, 1, 2) != "No") print(f)
#   # cat(f, "\n")
#   # cat(x, "\n\n")
# } # if nothng shows up, no variables were kicked because of collinarity issues
# 
# # look at VIF (if any are close to 3)
# for(f in VIF_files) {
#   x <- readLines(f)
#   x <- x[7:length(x)]
#   # if(substr(x, 1, 2) != "No") print(f)
#   cat("\n\n", f, "\n")
#    print(x)
# } # if nothng shows up, no variables were kicked because of collinarity issues
# 

# save all_Species_Year_to_keep to keep for Year analysis or climate interactions #####
if(solution_to_global_trend %in% "none") {
  write.csv(all_Species_Year_to_keep, file = "results/Species_Year_kept_for_Year_or_clim_x_dbh.csv", row.names = F)
  write.csv(species_removed_from_year_analysis, file = "results/Species_droped_for_Year_or_clim_x_dbh.csv", row.names = F )
}

} # for(solution_to_global_trend in ...)



# make an object holding the legend for species analyzed

all_legends <- list()
all_legends_2_cols <- list()
for(site in sites) {
  x <- all_Biol[[site]]
  # x <- x[!is.na(x$dbh),]
  # x <- x[x$species_code %in% summary_data$species_code[summary_data$site %in% site],]
  x <-  x[x$species_code %in% read.csv("results/summary_cores_analyzed.csv")$species_code[read.csv("results/summary_cores_analyzed.csv")$site %in% site],]
  x <- droplevels(x[!duplicated(x$species_code),])
  
  species_colors <- all_species_colors[[site]]
  names(species_colors) <- paste(x$genus, x$species)[match( names(species_colors), x$species_code)]

  a.gplot <- ggplot(x, aes(x = Year, y = dbh)) + geom_line(aes(group = paste(genus, species), col = paste(genus, species))) +  geom_ribbon(aes(ymin=min(dbh), ymax=max(dbh), bg = paste(genus, species)), alpha=0.25) + labs(col = sites_abb[[site]], bg = sites_abb[[site]])+ theme(legend.background = element_blank(), legend.box.background =element_blank(),  legend.justification = "left", legend.margin=margin(0,0,0,0), legend.box.margin=margin(0,0,0,10)) +  #+ guides(col=guide_legend(ncol=2), bg=guide_legend(ncol=2))
  scale_color_manual(values = species_colors[paste(x$genus, x$species)]) + scale_fill_manual(values = species_colors[paste(x$genus, x$species)])
  # a.gplot <- ggplot(x, aes(x = Year, y = dbh)) + geom_line(aes(group = paste0(genus, " ", species, " (", species_code, ")"), col = paste0(genus, " ", species, " (", species_code, ")"))) +  geom_ribbon(aes(ymin=min(dbh), ymax=max(dbh), bg = paste0(genus, " ", species, " (", species_code, ")")), alpha=0.25) + labs(col = sites_abb[[site]], bg = sites_abb[[site]])+ theme(legend.background = element_blank(), legend.box.background =element_blank()) #+ guides(col=guide_legend(ncol=2), bg=guide_legend(ncol=2))
  
  all_legends[[site]] <- ggplot_gtable(ggplot_build(a.gplot))$grobs[[which(sapply( ggplot_gtable(ggplot_build(a.gplot))$grobs, function(x) x$name) == "guide-box")]]
  
  a.gplot <- a.gplot + guides(col=guide_legend(ncol=2), bg=guide_legend(ncol=2))
  all_legends_2_cols[[site]]  <- ggplot_gtable(ggplot_build(a.gplot))$grobs[[which(sapply( ggplot_gtable(ggplot_build(a.gplot))$grobs, function(x) x$name) == "guide-box")]]
}

# make all legend the same width
maxwidth <- do.call(grid::unit.pmax, lapply(all_legends, function(x) x$widths[2:5]))
for (i in 1:length(all_legends)){
  all_legends[[i]]$widths[2:5] <- as.list(maxwidth)
}

save(list=c("all_legends", "all_legends_2_cols"), file = "results/all_legends.Rdata")

# summary plot for climate  ####

A <- pivot_longer(all_Clim, c(climate_variables, "SF"), "climate_var") # A <- pivot_longer(all_Clim, paste0(climate_variables, rep(c("", "_detrended"), each = length(climate_variables))), "climate_var") 

A$Year <- as.numeric(gsub("\\d\\d/\\d\\d/", "", A$Date))
A$value[A$value%in%-999] <- NA

A <- aggregate(value ~ sites_sitename + Year + climate_var, data = A, FUN =mean)
A <- rbind(A, data.frame(sites_sitename = "All",
                         Year = CO2$year,
                         climate_var = "CO2",
                         value = CO2$CO2_ppm))
A$group <- ifelse(grepl("_", A$climate_var), "dashed", "solid")

variables_units_full <- variables_units
variables_units_full[] <- paste(names(variables_units), variables_units, sep = "\n")

png("results/Climate_variables_yearly_mean.png", width = 8, height = 10, res = 300, units = "in")
ggplot(A[A$group == "solid",], aes(x = Year, y = value, color = sites_sitename)) +
  geom_line() + 
  facet_wrap(vars(gsub("_detrended", "", climate_var)), ncol =1, scales = "free_y", strip.position = "left", labeller = as_labeller(variables_units_full)) +
  ylab(NULL) +
  theme_classic() +
  theme(strip.background = element_blank(),
        strip.placement = "outside")
 
dev.off()
